fix(Promises): Fix signed and unsigned comparison in Pool

+++ Inc/C++Utilities/Pool.hpp
@@ -57,7 +57,7 @@ class Pool {
      * @note Could potentially release an element different than the original one if misused (eg. double free).
      */
     bool release(T* elem) {
-        if (elem < &elements[0] || elem - &elements[0] >= S) {
+        if (elem < &elements[0] || static_cast<size_t>(elem - &elements[0]) >= S) {
             return false;
         }
         size_t index = elem - &elements[0];
@@ -75,7 +75,7 @@ class Pool {
      * @return True if the element was successfully destroyed and released, false otherwise.
      */
     bool destroy(T* elem) {
-        if (elem < &elements[0] || elem - &elements[0] >= S) {
+        if (elem < &elements[0] || static_cast<size_t>(elem - &elements[0]) >= S) {
             return false;
         }
         size_t index = elem - &elements[0];
